#10: server action로직 분리
