--- src/utils/serverApi.ts
@@ -1,71 +1,0 @@
-"use server";
-
-import { BASE_URL, VERSIONS_API_URL } from "@/constants/url";
-import { ChampionDetailType, ChampionType } from "@/types/Champion";
-import { ItemType } from "@/types/Items";
-
-export const fetchVersionData = async (): Promise<string> => {
-  const versionResponse = await fetch(VERSIONS_API_URL, {
-    method: "GET",
-    cache: "force-cache",
-  });
-  const versionData: string[] = await versionResponse.json();
-  const currentVersion = versionData[0];
-
-  return currentVersion;
-};
-
-export const fetchItemsData = async (): Promise<{ ItemsData: ItemType }> => {
-  const currentVersion = await fetchVersionData();
-
-  try {
-    const ItemsResponse = await fetch(`${BASE_URL}/cdn/${currentVersion}/data/ko_KR/item.json`, {
-      method: "GET",
-      cache: "force-cache",
-    });
-    const ItemsData: ItemType = await ItemsResponse.json();
-
-    return { ItemsData };
-  } catch (error) {
-    console.error("아이템 데이터를 불러오는 중 오류가 발생했습니다.", error);
-    throw new Error("아이템 데이터를 불러오는 중 오류가 발생했습니다.");
-  }
-};
-
-export const fetchChampionsData = async (): Promise<{ championsData: ChampionType }> => {
-  const currentVersion = await fetchVersionData();
-
-  try {
-    const championsResponse = await fetch(`${BASE_URL}/cdn/${currentVersion}/data/ko_KR/champion.json`, {
-      method: "GET",
-      next: {
-        revalidate: 86400,
-      },
-    });
-
-    const championsData: ChampionType = await championsResponse.json();
-
-    return { championsData };
-  } catch (error) {
-    console.error("챔피언 데이터를 불러오는 중 오류가 발생했습니다.", error);
-    throw new Error("챔피언 데이터를 불러오는 중 오류가 발생했습니다.");
-  }
-};
-
-export const fetchChampionsDetailData = async (cid: string): Promise<{ championsDetailData: ChampionDetailType }> => {
-  const currentVersion = await fetchVersionData();
-
-  try {
-    const championsDetailResponse = await fetch(`${BASE_URL}/cdn/${currentVersion}/data/ko_KR/champion/${cid}.json`, {
-      method: "GET",
-      cache: "no-store",
-    });
-
-    const championsDetailData: ChampionDetailType = await championsDetailResponse.json();
-
-    return { championsDetailData };
-  } catch (error) {
-    console.error("챔피언 상세정보 데이터를 불러오는 중 오류가 발생했습니다.", error);
-    throw new Error("챔피언 상세정보 데이터를 불러오는 중 오류가 발생했습니다.");
-  }
-};
